Create 15.Next Day
Write a JS function that calculates the date of the next day by given year, month and day.
The input comes as three number parameters. The first element is the year, the second is the month and the third is the day.
The output should be returned as a result of your function.